Add files via upload
Open research institute (ORI), is a committed to open source work ranging in various fields. One of its flagship project is P4G: Phase4Ground, an open source HAM satellite, being built to address the needs of  Hobbyists, facilitate emergency communications etc. 

This Presentation, is a poster designed (for the proposed AAAAA protocol) which is displayed at the DEFCON 2022 conference.